Apostrophes, "in retrospect".
> it's designed principles
> it's validity is enforced.

These aren't a contraction of "it is" or "it has", so no apostrophe. It's hella confusing because one does write "I like Rust's style" with an apostrophe for the possessive form, but "its" is already the possessive pronoun. The way I remember it is: you don't write "I like hi's principles," and "its principles" and even "the idea was hers" work the same.

The phrase is "obvious in retrospect", I also tweaked it to be, arguably, more idiomatic.

Your enthusiasm, even love, for the Rust language comes across clearly and that's what really matters :heart_eyes: