use font-size-adjust
This is radically great. The root of many problems in web design is that
many font metrics have arbitray base units. Two fonts at font-size 16px
can look _widely_ different in size, because there's no useful defintion
for the size.

16px is the size of "canvans" onto a letter is drawn by the font
designer (originating from a literal lead rectange of a physical font)
but canvas can be big or small. See https://tonsky.me/blog/font-size/
for details.

What `font-size-adjust` allows is to look at a specific metric, like the
height of the "x" gliph, and harmonize fonts along _that_.

I have there fonts, Open Sans, Garamond, and JetBrains Mono, and they
are very different in terms of font-size, which required a bunch of
override. But with font-size-adjust, I no longer need to set their
_relative_ font sizes.

The constant 0.53 is somewhat arbitray, I picked it as a rough fixed
point for Open Sans and Helvetica.

Kudos to xoria.org for showing me the light (which I didn't see, of
course, for I am blind, but that is a different story...)